fix if statements, fix object construction

+++ common/client/recordings.cs
@@ -5,12 +5,14 @@ function recordingsDlg::onWake() {
     %filespec = $currentMod @ "/recordings/*.rec";
     echo(%filespec);
     %file = findFirstFile(%filespec);
-    %fileName = fileBase(%file);
-    !((%file $= ""));
-    %i = (1.0 + %i);
-    RecordingsDlgList.addRow(%fileName);
-    %file = findNextFile(%filespec);
-    (-(1.0) == strstr(%file, "/CVS/"));
+    if (!(%file $= "")) {
+        %fileName = fileBase(%file);
+        if ((-(1.0) == strstr(%file, "/CVS/"))) {
+            %i = (1.0 + %i);
+            RecordingsDlgList.addRow(%fileName);
+        }
+        %file = findNextFile(%filespec);
+    }
     0.sort();
     0.setSelectedRow();
     0.scrollVisible();
@@ -21,39 +23,57 @@ function StartSelectedDemo() {
     %rowText = %sel.getRowTextById();
     RecordingsDlgList;
     %file = $currentMod @ "/recordings/" @ getField(%rowText, 0) @ ".rec";
-    new ();
+    new GameConnection(ServerConnection);
     "".setCommonPreconnectClientSettings();
     add();
-    setContent();
-    popDialog();
-    prepDemoPlayback();
+    if (%file.playDemo()) {
+        setContent();
+        popDialog();
+        prepDemoPlayback();
+    }
     MessageBoxOK("Playback Failed", recordingsDlg @ ServerConnection @ "Demo playback failed for file '" @ %file @ "'.", "");
-    delete();
+    if (isObject()) {
+        delete();
+    }
 };
 function startDemoRecord() {
     stopRecording();
-    return isDemoPlaying();
+    if (isDemoPlaying()) {
+        return ServerConnection;
+    }
     %i = 0;
-    %num = %i;
-    (1000.0 < %i);
-    %num = (10.0 < %num) @ 0 @ %num;
-    %num = (100.0 < %num) @ 0 @ %num;
-    %file = $currentMod @ "/recordings/demo" @ %num @ ".rec";
-    %i = (1.0 + %i);
-    !(isFile(%file));
-    return (1000.0 == %i);
+    if ((1000.0 < %i)) {
+        %num = %i;
+        if ((10.0 < %num)) {
+            %num = 0 @ %num;
+        }
+        if ((100.0 < %num)) {
+            %num = 0 @ %num;
+        }
+        %file = $currentMod @ "/recordings/demo" @ %num @ ".rec";
+        if (!(isFile(%file))) {
+        }
+        %i = (1.0 + %i);
+    }
+    if ((1000.0 == %i)) {
+        return (1000.0 < %i);
+    }
     $DemoFileName = %file;
     ChatHud @ "\x05Recording to file [\x03" @ $DemoFileName @ "\x0F].".addLine();
     prepDemoRecord();
     $DemoFileName.startRecording();
-    deleteFile($DemoFileName);
-    !(isDemoRecording()) @ ChatHud @ "\x04 *** Failed to record to file [\x03" @ $DemoFileName @ "\x0F].".addLine();
-    $DemoFileName = "";
-    ServerConnection;
+    if (!(isDemoRecording())) {
+        deleteFile($DemoFileName);
+        ServerConnection @ ChatHud @ "\x04 *** Failed to record to file [\x03" @ $DemoFileName @ "\x0F].".addLine();
+        $DemoFileName = "";
+        ServerConnection;
+    }
 };
 function stopDemoRecord() {
-    isDemoRecording() @ ChatHud @ "\x05Recording file [\x03" @ $DemoFileName @ "\x0F] finished.".addLine();
-    stopRecording();
+    if (isDemoRecording()) {
+        ServerConnection @ ChatHud @ "\x05Recording file [\x03" @ $DemoFileName @ "\x0F] finished.".addLine();
+        stopRecording();
+    }
 };
 function demoPlaybackComplete() {
     disconnect();
