+++ common/server/server.cs
@@ -1,39 +1,40 @@
 exec("./dif2dae.cs");
 function portInit(%port) {
     %failCount = 0;
-    echo(!(setNetPort(%port)) @ "Port init failed on port " @ %port @ " trying next port.");
-    %port = (1.0 + %port);
-    %failCount = (1.0 + %failCount);
+    if (!(setNetPort(%port))) {
+        echo("Port init failed on port " @ %port @ " trying next port.");
+        %port = (1.0 + %port);
+        %failCount = (1.0 + %failCount);
+    }
     $Net::BoundPort = %port;
     !(setNetPort(%port));
     return %failCount;
 };
 function createServer(%serverType, %mission) {
-    error("createServer: mission name unspecified");
-    return (%mission $= "");
+    if ((%mission $= "")) {
+        error("createServer: mission name unspecified");
+        return;
+    }
     destroyServer();
     $MissionSequence = 0;
     $Server::ServerType = %serverType;
     $Net::BoundPort = 0;
-    portInit($Pref::Server::Port);
-    allowConnections(1);
-    $ServerGroup = new ();
-    ServerGroup;
-    $ClientDict = new ();
-    ClientDict;
-    $PlayerDict = new ();
-    PlayerDict;
-    $TokenDict = new ();
-    TokenDict;
-    $PendingValidate = new ();
-    PendingValidate;
+    if ((%serverType $= "MultiPlayer")) {
+        portInit($Pref::Server::Port);
+        allowConnections(1);
+    }
+    $ServerGroup = new SimGroup(ServerGroup);
+    $ClientDict = new StringMap(ClientDict);
+    $PlayerDict = new StringMap(PlayerDict);
+    $TokenDict = new StringMap(TokenDict);
+    $PendingValidate = new StringMap(PendingValidate);
     allowInstanceMethods();
     allowInstanceMethods();
     allowInstanceMethods();
-    new ();
+    new StringMap(PlayerNameLowerToRegMap);
     onServerCreated();
     loadMission(%mission, 1);
-    return PlayerNameLowerToRegMap;
+    return TokenDict;
 };
 function destroyServer() {
     $Server::ServerType = "";
@@ -41,12 +42,20 @@ function destroyServer() {
     $missionRunning = 0;
     endMission();
     onServerDestroyed();
-    delete();
-    delete();
-    $ServerGroup.delete();
-    %client = 0.getObject();
-    ClientGroup;
-    %client.delete();
+    if (isObject()) {
+        delete();
+    }
+    if (isObject()) {
+        delete();
+    }
+    if (isObject($ServerGroup)) {
+        $ServerGroup.delete();
+    }
+    if (getCount()) {
+        %client = 0.getObject();
+        ClientGroup;
+        %client.delete();
+    }
     $Server::GuidList = "";
     getCount();
     deleteDataBlocks();
@@ -63,24 +72,35 @@ function resetServerDefaults() {
 function addToServerGuidList(%guid) {
     %count = getFieldCount($Server::GuidList);
     %i = 0;
-    return (%guid == getField($Server::GuidList, %i));
-    %i = (1.0 + %i);
+    if ((%count < %i)) {
+        if ((%guid == getField($Server::GuidList, %i))) {
+            return;
+        }
+        %i = (1.0 + %i);
+    }
+    if (((%count < %i) SPC $Server::GuidList $= "")) {
+    }
     $Server::GuidList = $Server::GuidList;
     %guid;
-    return ((%count < %i) SPC $Server::GuidList $= "");
+    return;
 };
 function removeFromServerGuidList(%guid) {
     %count = getFieldCount($Server::GuidList);
     %i = 0;
-    $Server::GuidList = removeField($Server::GuidList, %i);
-    (%guid == getField($Server::GuidList, %i));
-    return (%count < %i);
-    %i = (1.0 + %i);
+    if ((%count < %i)) {
+        if ((%guid == getField($Server::GuidList, %i))) {
+            $Server::GuidList = removeField($Server::GuidList, %i);
+            return;
+        }
+        %i = (1.0 + %i);
+    }
 };
 function isUserConnected(%userName) {
     %client = %userName.getNorm();
     ClientDict;
-    return 1;
+    if (!(%client $= "")) {
+        return 1;
+    }
     return 0;
 };
 function onServerInfoQuery() {
